requirements.txt
Flask
flask-bootstrap
numpy
Pillow
opencv-python
gunicorn